requirements.txt
# requirements.txt
# Основные зависимости для стенда машинного зрения

flask>=2.0.0          # Веб-интерфейс
pymodbus>=3.0.0       # Modbus TCP для связи с ОВЕН и камерой
pyyaml>=5.4.0         # Чтение конфигурации YAML
waitress>=2.1.0       # Производственный WSGI-сервер (замена встроенному Flask)

# Для работы с Excel (таблица инструментов)
pandas>=2.0.0         # Обработка данных
openpyxl>=3.0.0       # Чтение Excel-файлов


psycopg2-binary>=2.9.0
# Для работы с изображениями (если потребуется сохранять/обрабатывать снимки)
# opencv-python>=4.5.0

# Примечание: все зависимости устанавливаются через pip install -r requirements.txt